Split out SubmitApplication choice service from SubmitApplication
